Enum tor_dirmgr::Error

#[non_exhaustive]
pub enum Error {

    Unwanted(&'static str),
    NetDirOlder,
    NoDownloadSupport,
    CacheCorruption(&'static str),
    SqliteError(Arc<Error>),
    UnrecognizedSchema {
        schema: u32,
        supported: u32,
    },
    DirectoryNotPresent,
    UnrecognizedAuthorities,
    ManagerDropped,
    CantAdvanceState,
    LockFile(Arc<Error>),
    CacheFile {
        action: &'static str,
        fname: PathBuf,
        error: Arc<Error>,
    },
    ConsensusDiffError(Error),
    BadUtf8FromDirectory(FromUtf8Error),
    BadUtf8InCache(Utf8Error),
    BadHexInCache(FromHexError),
    NetDocError {
        source: DocSource,
        cause: Error,
    },
    ConsensusInvalid {
        source: DocSource,
        cause: Error,
    },
    UntimelyObject(TimeValidityError),
    DirClientError(Error),
    SignatureError(Arc<Error>),
    OfflineMode,
    CachePermissions(Error),
    Spawn {
        spawning: &'static str,
        cause: Arc<SpawnError>,
    },
    ExternalDirProvider {
        cause: Arc<dyn Error + Send + Sync + 'static>,
        kind: ErrorKind,
    },
    Bug(Bug),

}

An error originated by the directory manager code

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Unwanted(&'static str)

We received a document we didn’t want at all.

NetDirOlder

The NetDir we downloaded is older than the one we already have.

NoDownloadSupport

This DirMgr doesn’t support downloads.

CacheCorruption(&'static str)

We couldn’t read something from disk that we should have been able to read.

SqliteError(Arc<Error>)

rusqlite gave us an error.

UnrecognizedSchema

Fields

schema: u32

The schema version in the database

supported: u32

The schema that we actually support.

A schema version that says we can’t read it.

DirectoryNotPresent

User requested an operation that required a usable bootstrapped directory, but we didn’t have one.

UnrecognizedAuthorities

A consensus document is signed by an unrecognized authority set.

ManagerDropped

A directory manager has been dropped; background tasks can exit too.

CantAdvanceState

We made a bunch of attempts, but weren’t unable to advance the state of a download.

LockFile(Arc<Error>)

Error while accessing a lockfile.

CacheFile

Fields

action: &'static str

What we were doing when we encountered the error.

fname: PathBuf

The file that we were trying to access.

error: Arc<Error>

The underlying IO error.

Error while accessing a file in the store.

ConsensusDiffError(Error)

An error given by the consensus diff crate.

BadUtf8FromDirectory(FromUtf8Error)

Invalid UTF8 in directory response.

BadUtf8InCache(Utf8Error)

Invalid UTF8 from our cache.

BadHexInCache(FromHexError)

Invalid hexadecimal value in the cache.

NetDocError

Fields

source: DocSource

Where the document came from.

cause: Error

What error we got.

An error given by the network document crate.

ConsensusInvalid

Fields

source: DocSource

Where the document came from.

cause: Error

What error we got.

An error indicating that the consensus could not be validated.

This kind of error is only returned during the certificate fetching state; it indicates that a consensus which previously seemed to be plausible has turned out to be wrong after we got the certificates.

UntimelyObject(TimeValidityError)

An error caused by an expired or not-yet-valid object.

DirClientError(Error)

An error given by dirclient

SignatureError(Arc<Error>)

An error given by the checkable crate.

OfflineMode

An attempt was made to bootstrap a DirMgr created in offline mode.

CachePermissions(Error)

A problem with file permissions on our cache directory.

Spawn

Fields

spawning: &'static str

What we were trying to spawn

cause: Arc<SpawnError>

What happened when we tried to spawn it

Unable to spawn task

ExternalDirProvider

Fields

cause: Arc<dyn Error + Send + Sync + 'static>

What happened

kind: ErrorKind

The kind

Other error from an external directory provider

Bug(Bug)

A programming problem, either in our code or the code calling it.