Struct tor_dirmgr::DirMgr

pub struct DirMgr<R: Runtime> { /* private fields */ }

A directory manager to download, fetch, and cache a Tor directory.

A DirMgr can operate in three modes:

In offline mode, it only reads from the cache, and can only read once.

In read-only mode, it reads from the cache, but checks whether it can acquire an associated lock file. If it can, then it enters read-write mode. If not, it checks the cache periodically for new information.

In read-write mode, it knows that no other process will be writing to the cache, and it takes responsibility for fetching data from the network and updating the directory with new directory information.

Implementations

impl<R: Runtime> DirMgr<R>

pub async fn load_once(runtime: R, config: DirMgrConfig) -> Result<Arc<NetDir>>

Try to load the directory from disk, without launching any kind of update process.

This function runs in offline mode: it will give an error if the result is not up-to-date, or not fully downloaded.

In general, you shouldn’t use this function in a long-running program; it’s only suitable for command-line or batch tools.

pub async fn load_or_bootstrap_once( config: DirMgrConfig, runtime: R, store: DirMgrStore<R>, circmgr: Arc<CircMgr<R>> ) -> Result<Arc<NetDir>>

Return a current netdir, either loading it or bootstrapping it as needed.

Like load_once, but will try to bootstrap (or wait for another process to bootstrap) if we don’t have an up-to-date bootstrapped directory.

In general, you shouldn’t use this function in a long-running program; it’s only suitable for command-line or batch tools.

pub fn create_unbootstrapped( config: DirMgrConfig, runtime: R, store: DirMgrStore<R>, circmgr: Arc<CircMgr<R>> ) -> Result<Arc<Self>>

Create a new DirMgr in online mode, but don’t bootstrap it yet.

The DirMgr can be bootstrapped later with bootstrap.

pub async fn bootstrap(self: &Arc<Self>) -> Result<()>

Bootstrap a DirMgr created in online mode that hasn’t been bootstrapped yet.

This function will not return until the directory is bootstrapped enough to build circuits. It will also launch a background task that fetches any missing information, and that replaces the directory when a new one is available.

This function is intended to be used together with create_unbootstrapped. There is no need to call this function otherwise.

If bootstrapping has already successfully taken place, returns early with success.

Errors

Returns an error if bootstrapping fails. If the error is Error::CantAdvanceState, it may be possible to successfully bootstrap later on by calling this function again.

Panics

Panics if the DirMgr passed to this function was not created in online mode, such as via load_once.

pub fn bootstrap_started(&self) -> bool

Returns true if a bootstrap attempt is in progress, or successfully completed.

pub async fn bootstrap_from_config( config: DirMgrConfig, runtime: R, store: DirMgrStore<R>, circmgr: Arc<CircMgr<R>> ) -> Result<Arc<Self>>

Return a new directory manager from a given configuration, bootstrapping from the network as necessary.

pub fn reconfigure( &self, new_config: &DirMgrConfig, how: Reconfigure ) -> Result<(), ReconfigureError>

Try to change our configuration to new_config.

Actual behavior will depend on the value of how.

pub fn bootstrap_events(&self) -> DirBootstrapEvents

Return a stream of DirBootstrapStatus events to tell us about changes in the latest directory’s bootstrap status.

Note that this stream can be lossy: the caller will not necessarily observe every event on the stream

pub fn events(&self) -> impl Stream<Item = DirEvent>

Return a new asynchronous stream that will receive notification whenever the consensus has changed.

Multiple events may be batched up into a single item: each time this stream yields an event, all you can assume is that the event has occurred at least once.

pub fn text(&self, doc: &DocId) -> Result<Option<DocumentText>>

Try to load the text of a single document described by doc from storage.

pub fn texts<T>(&self, docs: T) -> Result<HashMap<DocId, DocumentText>>where T: IntoIterator<Item = DocId>,

Load the text for a collection of documents.

If many of the documents have the same type, this can be more efficient than calling text.